requirements.txt
numpy
PyQt5
keras
pillow
opencv-python